Часть 2.1

+++ main-service/src/main/java/ru/yandex/practicum/event/UpdateEventUserRequest.java
@@ -15,7 +15,7 @@ import java.time.LocalDateTime;
 public class UpdateEventUserRequest {
     @Size(min = 20, max = 2000)
     private String annotation;
-    private Long category;
+    private Integer category;
     @Size(min = 20, max = 7000)
     private String description;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
